-renamed chat view footer to text entry footer -moved text entry footer to core widgets
