+++ lib/core/widgets/text entry footer/chat_send_button.dart
