+++ lib/core/widgets/text entry footer/chat_text_field.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:grad_project/core/helpers/constants.dart';
 import 'package:grad_project/core/theme/app_text_styles.dart';
-import 'package:grad_project/features/chat/ui/widgets/chat%20view%20footer/mention_icon_button.dart';
+import 'package:grad_project/core/widgets/text%20entry%20footer/mention_icon_button.dart';
 import 'package:intl/intl.dart';
-import '../../../../../core/theme/app_colors.dart';
-import '../../../../../generated/l10n.dart';
+import '../../theme/app_colors.dart';
+import '../../../generated/l10n.dart';
 import 'emoji_picker_icon_button.dart';
 
 class ChatTextField extends StatelessWidget {
