+++ lib/core/widgets/text entry footer/emoji_picker_icon_button.dart
